Rubiks cube class files.
This a Rubiks cube game built with processing(Java). I have looked all over the internet, and no one has entirely done this. This is a rubiks cube that you can play and solve on the computer (except for the actual physical rubiks cube in your hand) built with processing because of its friendly ecosystem for graphics. When you start the application/ game, the rubiks cube will be shuffled, and you will be able to move around the cubes with input from the keyboard with the keys: l, r, f, b, d, u, i, e, g, v, k, s, y. With these keys you will be able to move the cubes both clock-wise and anti-clock-wise. Once you have solved the cube, a dialog box will show saying you have completed solving the cube. There is a landing gui for the user before beginning solving the Rubiks cube. More functionality i want to add to the application is to have levels, once you have solved the rubiks cube with 3 by 3 by 3 cubes, move on to solve rubiks cube with 4 by 4 by 4 cubes.